+++ Rdmp.UI/CommandExecution/AtomicCommands/UIFactory/AtomicCommandUIFactory.cs
@@ -30,12 +30,12 @@ namespace Rdmp.UI.CommandExecution.AtomicCommands.UIFactory
 
         public ToolStripMenuItem CreateMenuItem(IAtomicCommand command)
         {
-            return new AtomicCommandMenuItem(command, _activator);
+            return new AtomicCommandMenuItem(command, _activator){Tag = command };
         }
 
         public AtomicCommandLinkLabel CreateLinkLabel(IAtomicCommand command)
         {
-            return new AtomicCommandLinkLabel(_iconProvider,command);
+            return new AtomicCommandLinkLabel(_iconProvider,command){Tag = command };
         }
 
         public RDMPContextMenuStrip CreateMenu(IActivateItems activator,TreeListView tree, RDMPCollection collection, params IAtomicCommand[] commands)