+++ Tools/rdmp/CommandLine/Gui/ConsoleGuiSqlEditor.cs
@@ -4,13 +4,19 @@
 // RDMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
+using CsvHelper;
 using Rdmp.Core.CommandExecution;
+using Rdmp.Core.DataExport.DataExtraction;
 using Rdmp.Core.DataViewing;
 using ReusableLibraryCode.DataAccess;
 using System;
 using System.Collections.Generic;
 using System.Data;
+using System.Data.Common;
+using System.Globalization;
+using System.IO;
 using System.Text;
+using System.Threading.Tasks;
 using Terminal.Gui;
 
 namespace Rdmp.Core.CommandLine.Gui
@@ -21,6 +27,24 @@ namespace Rdmp.Core.CommandLine.Gui
         private readonly IViewSQLAndResultsCollection _collection;
         private TableView tableView;
         private TextView textView;
+        private Button _btnRunOrCancel;
+        private Task _runSqlTask;
+        private DbCommand _runSqlCmd;
+
+        /// <summary>
+        /// The original SQL this control was launched with
+        /// </summary>
+        private string _orignalSql;
+
+        /// <summary>
+        /// The number of seconds to allow queries to run for, can be changed by user
+        /// </summary>
+        private int _timeout = DefaultTimeout;
+
+        /// <summary>
+        /// The default number of seconds to allow queries to run for when no value or an invalid value is specified by the user
+        /// </summary>
+        public const int DefaultTimeout = 300;
 
         public ConsoleGuiSqlEditor(IBasicActivateItems activator,IViewSQLAndResultsCollection collection)
         {
@@ -34,21 +58,61 @@ namespace Rdmp.Core.CommandLine.Gui
                 Y=0,
                 Width = Dim.Fill(),
                 Height = Dim.Percent(30),
-                Text = collection.GetSql().Replace("\r\n","\n")
+                Text = _orignalSql = collection.GetSql().Replace("\r\n","\n")
             };
 
             Add(textView);
 
-            var btnRun = new Button("Run"){
+            _btnRunOrCancel = new Button("Run"){
                 X= 0,
                 Y=Pos.Bottom(textView),
                 };
 
-            btnRun.Clicked += ()=>RunSql();
-            Add(btnRun);
+            _btnRunOrCancel.Clicked += ()=>RunOrCancel();
+            Add(_btnRunOrCancel);
+
+            var resetSql = new Button("Reset Sql"){
+                X= Pos.Right(_btnRunOrCancel)+1,
+                Y= Pos.Bottom(textView),
+                };
+
+            resetSql.Clicked += ()=>ResetSql();
+            Add(resetSql);
+
+            var clearSql = new Button("Clear Sql"){
+                X= Pos.Right(resetSql)+1,
+                Y= Pos.Bottom(textView),
+                };
+
+            clearSql.Clicked += ()=>ClearSql();
+            Add(clearSql);
+
+            var lblTimeout = new Label("Timeout:")
+            {
+                X = Pos.Right(clearSql)+1,
+                Y = Pos.Bottom(textView)
+            };
+            Add(lblTimeout);
+
+            var tbTimeout = new TextField(_timeout.ToString())
+            {
+                X = Pos.Right(lblTimeout),
+                Y = Pos.Bottom(textView),
+                Width = 5
+            };
+            tbTimeout.TextChanged += TbTimeout_TextChanged;
+            
+            Add(tbTimeout);
+
+            var btnSave = new Button("Save"){
+                X= Pos.Right(tbTimeout)+1,
+                Y= Pos.Bottom(textView),
+                };
+            btnSave.Clicked += ()=>Save();
+            Add(btnSave);
 
             var btnClose = new Button("Close"){
-                X= Pos.Right(btnRun),
+                X= Pos.Right(btnSave)+1,
                 Y= Pos.Bottom(textView),
                 };
             btnClose.Clicked += ()=>Application.RequestStop();
@@ -65,6 +129,98 @@ namespace Rdmp.Core.CommandLine.Gui
             Add(tableView);
         }
 
+        private void Save()
+        {
+            try
+            {
+                var tbl = tableView.Table;
+
+                if(tbl == null)
+                {
+                    MessageBox.ErrorQuery("Cannot Save","No Table Loaded","Ok");
+                    return;
+                }
+
+                var sfd = new SaveDialog("Save","Pick file location to save");
+                Application.Run(sfd);
+
+                if(sfd.Canceled)
+                    return;
+
+                if(sfd.FilePath != null)
+                {
+                    using(var writer = new StreamWriter(File.OpenWrite(sfd.FilePath.ToString())))
+                        using(var w = new CsvWriter(writer,CultureInfo.CurrentCulture))
+                        {
+                            // write headers
+                            foreach(DataColumn c in tbl.Columns)
+                                w.WriteField(c.ColumnName);
+
+                            w.NextRecord();
+
+                            // write rows
+                            foreach (DataRow r in tbl.Rows)
+                            {
+                                foreach (var item in r.ItemArray)
+                                {
+                                    w.WriteField(item);
+                                }
+
+                                w.NextRecord();
+                            }
+                        }
+                    
+                    MessageBox.Query("File Saved","Save completed","Ok");
+                }
+            }
+            catch (Exception ex)
+            {
+                MessageBox.ErrorQuery("Save Failed",ex.Message,"Ok");
+            }
+            
+        }
+
+        private void TbTimeout_TextChanged(NStack.ustring value)
+        {
+            if(int.TryParse(value.ToString(),out int newTimeout))
+                _timeout = newTimeout < 0 ? DefaultTimeout : newTimeout;
+            else
+                _timeout = DefaultTimeout;
+        }
+
+        private void ClearSql()
+        {
+            textView.Text = "";
+            textView.SetNeedsDisplay();
+        }
+
+        private void ResetSql()
+        {
+            textView.Text = _orignalSql;
+            textView.SetNeedsDisplay();
+        }
+
+        private void RunOrCancel()
+        {
+            // if task is still running we should cancel
+            if(_runSqlTask  != null && !_runSqlTask.IsCompleted)
+            {
+                // Cancel the sql command and let that naturally end the task
+                _runSqlCmd?.Cancel();
+            }
+            else
+            {
+                _runSqlTask = Task.Run(()=>RunSql());
+                _btnRunOrCancel.Text = "Cancel";
+                _btnRunOrCancel.SetNeedsDisplay();
+            }
+        }
+        
+        private void SetReadyToRun()
+        {
+            _btnRunOrCancel.Text = "Run";
+            _btnRunOrCancel.SetNeedsDisplay();
+        }
         private void RunSql()
         {
             try
@@ -82,7 +238,10 @@ namespace Rdmp.Core.CommandLine.Gui
                 using(var con = db.Server.GetConnection())
                 {
                     con.Open();
-                    using(var da = db.Server.GetDataAdapter(sql,con))
+                    _runSqlCmd = db.Server.GetCommand(sql,con);
+                    _runSqlCmd.CommandTimeout = _timeout;
+
+                    using(var da = db.Server.GetDataAdapter(_runSqlCmd))
                     {
                         var dt = new DataTable();
                         da.Fill(dt);
@@ -94,6 +253,10 @@ namespace Rdmp.Core.CommandLine.Gui
             catch (Exception ex)
             {
                 _activator.ShowException("Failed to run query",ex);
+            }
+            finally
+            {
+                SetReadyToRun();
             }
         }
     }